docker_compose/simple_project/requirements.txt
asgiref==3.7.2
dataclasses==0.6
Django==4.2.5
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uWSGI==2.0.24